+++ examples/whip/src/main.rs
@@ -103,7 +103,6 @@ fn i420_planes_to_buffer(
 pub struct FactoryHolder {
     factory: PeerConnectionFactory,
     _network: Thread,
-    _worker: Thread,
     _signaling: Thread,
 }
 
@@ -111,15 +110,13 @@ impl FactoryHolder {
     pub fn new() -> Option<Arc<Self>> {
         let env = Environment::new();
         let mut network = Thread::new_with_socket_server();
-        let mut worker = Thread::new();
         let mut signaling = Thread::new();
         network.start();
-        worker.start();
         signaling.start();
 
         let mut deps = PeerConnectionFactoryDependencies::new();
         deps.set_network_thread(&network);
-        deps.set_worker_thread(&worker);
+        deps.set_worker_thread(&network);
         deps.set_signaling_thread(&signaling);
         let event_log = RtcEventLogFactory::new();
         deps.set_event_log_factory(event_log);
@@ -142,7 +139,6 @@ impl FactoryHolder {
         Some(Arc::new(Self {
             factory,
             _network: network,
-            _worker: worker,
             _signaling: signaling,
         }))
     }