+++ webrtc/src/whep.c
@@ -29,7 +29,6 @@ static void whep_OnSendRequestResponse(char* resp, void* user_data);
 struct PeerConnectionFactory {
   struct webrtc_RefCountInterface_ref* ref;
   struct webrtc_Thread_unique* network_thread;
-  struct webrtc_Thread_unique* worker_thread;
   struct webrtc_Thread_unique* signaling_thread;
   struct webrtc_PeerConnectionFactoryInterface_refcounted* factory;
 };
@@ -43,10 +42,6 @@ void PeerConnectionFactory_delete(void* user_data) {
   if (p->network_thread != NULL) {
     webrtc_Thread_Stop(webrtc_Thread_unique_get(p->network_thread));
     webrtc_Thread_unique_delete(p->network_thread);
-  }
-  if (p->worker_thread != NULL) {
-    webrtc_Thread_Stop(webrtc_Thread_unique_get(p->worker_thread));
-    webrtc_Thread_unique_delete(p->worker_thread);
   }
   if (p->signaling_thread != NULL) {
     webrtc_Thread_Stop(webrtc_Thread_unique_get(p->signaling_thread));
@@ -345,8 +340,6 @@ struct PeerConnectionFactory* PeerConnectionFactory_Create() {
 
   p->network_thread = webrtc_Thread_CreateWithSocketServer();
   webrtc_Thread_Start(webrtc_Thread_unique_get(p->network_thread));
-  p->worker_thread = webrtc_Thread_Create();
-  webrtc_Thread_Start(webrtc_Thread_unique_get(p->worker_thread));
   p->signaling_thread = webrtc_Thread_Create();
   webrtc_Thread_Start(webrtc_Thread_unique_get(p->signaling_thread));
 
@@ -356,14 +349,14 @@ struct PeerConnectionFactory* PeerConnectionFactory_Create() {
   webrtc_PeerConnectionFactoryDependencies_set_network_thread(
       dependencies, webrtc_Thread_unique_get(p->network_thread));
   webrtc_PeerConnectionFactoryDependencies_set_worker_thread(
-      dependencies, webrtc_Thread_unique_get(p->worker_thread));
+      dependencies, webrtc_Thread_unique_get(p->network_thread));
   webrtc_PeerConnectionFactoryDependencies_set_signaling_thread(
       dependencies, webrtc_Thread_unique_get(p->signaling_thread));
   webrtc_PeerConnectionFactoryDependencies_set_event_log_factory(
       dependencies, webrtc_RtcEventLogFactory_Create());
   struct webrtc_AudioDeviceModule_refcounted* adm =
       (struct webrtc_AudioDeviceModule_refcounted*)webrtc_Thread_BlockingCall_r(
-          webrtc_Thread_unique_get(p->worker_thread), _BlockingCall_create_adm,
+          webrtc_Thread_unique_get(p->network_thread), _BlockingCall_create_adm,
           env);
   webrtc_PeerConnectionFactoryDependencies_set_adm(dependencies, adm);
   webrtc_AudioDeviceModule_Release(